backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
web3==6.15.1
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
requests==2.31.0
pymongo==4.6.0
cryptography==42.0.5
solcx==1.1.1
